Kit Menu
Kit Menu and Purchased Menu rework. Changed KitMenu to constructor. Created Purchased Menu class to open from KitMenu
