+++ src/net/peacefulcraft/trenchpvp/gameclasses/abilities/BigBertha.java
@@ -17,6 +17,9 @@ import org.bukkit.util.Vector;
 
 import net.peacefulcraft.trenchpvp.TrenchPvP;
 import net.peacefulcraft.trenchpvp.gameclasses.classConfigurations.TrenchKit;
+import net.peacefulcraft.trenchpvp.gameclasses.classConfigurations.TrenchKits;
+import net.peacefulcraft.trenchpvp.gamehandle.TeamManager;
+import net.peacefulcraft.trenchpvp.gamehandle.player.TrenchPlayer;
 import net.peacefulcraft.trenchpvp.stats.TrenchStats.DemoStat;
 
 public class BigBertha extends TrenchAbility{
@@ -66,6 +69,9 @@ public class BigBertha extends TrenchAbility{
 				//Schedule giving player another Bertha 25 seconds from now
 				Bukkit.getServer().getScheduler().scheduleSyncDelayedTask(TrenchPvP.getPluginInstance() , new Runnable() {
 	                public void run() {
+	                	TrenchPlayer t = TeamManager.findTrenchPlayer(p);
+	                	if(!(t.getKitType() != TrenchKits.DEMOMAN)) {return;}
+	                		                	
 	                	p.getInventory().setItem(itemIndex, bomb);
 	                }
 	            }, 300);
