+++ src/net/peacefulcraft/trenchpvp/gameclasses/classConfigurations/TrenchKits.java
@@ -4,5 +4,5 @@ public enum TrenchKits {
 	UNASSIGNED,SCOUT,SOLDIER,PYRO,
 	DEMOMAN,HEAVY,SNIPER,MEDIC,SPY,
 	
-	ADRENALINE_JUNKY;
+	ADRENALINE_JUNKIE;
 }
